Add function to handle static pages
This change refactors the application to add a function that can handle
static pages (pages with no dynamic content). The intention of the
change was twofold. Firstly, I wanted to learn more about Go in a
hands-on practical way. Secondly, I wanted to add the next aspect of
functionality to the application from the PHP/Python version, that being
rendering static pages. I don't know if this is the best approach, but
it does the job. Plus, I learned a little bit about a different router:
Gorilla Mux.