Refactor core declaration of template into a function
I don't know if this is the best approach, nor an acceptable one, but,
conceptually, removing duplicate code is a good thing, so I'm okay with
this change for that reason. The intent is to do that and provide one
place where the templates are setup, before being used. The change also
moves explicit declaration of the data structs into the call to the
template execution. I just wanted to see if this could be done and if it
seemed to be a more streamlined way of doing things.